space shooter/requirements.txt
nump==5.5.5.5
numpy==2.2.3
pygame-ce==2.5.3
